+++ src/defines.h
@@ -13,7 +13,7 @@
 //-------------------------------------
 #define VERSION_MAJOR       0
 #define VERSION_MINOR       7
-#define VERSION_PATCH       62
+#define VERSION_PATCH       63
 
 //-------------------------------------
 typedef struct {